# contracts/openapi.yaml
openapi: 3.1.0
...
info:
  title: Accounting Platform API
  version: 1.0.0
  description: |
    Multi-tenant accounting platform. All business resources are tenant-scoped and most are
    legal-entity scoped; scope is taken from the authenticated session, never from a request
    parameter. Money is carried as decimal strings, never JSON numbers (ADR-0006).
  x-supersedes:
    - project plan/blueprints/api_route_catalog.yaml
servers:
  - url: https://api.example.com/v1
    description: Production
  - url: https://sandbox.api.example.com/v1
    description: Sandbox (doc 15 developer platform)
security:
  - sessionBearer: []
  - oauth2: []
...
paths:
...
  /approval-workflows:
    get:  {operationId: listApprovalWorkflows,  tags: [Workflow], x-permission: workflow.view,   x-idempotency: n-a,         x-phase: 6, responses: {'200': {$ref: '#/components/responses/Ok'},      default: {$ref: '#/components/responses/Error'}}}
    post: {operationId: createApprovalWorkflow, tags: [Workflow], x-permission: workflow.manage, x-idempotency: recommended, x-phase: 6, x-supersedes: "/workflows", responses: {'201': {$ref: '#/components/responses/Created'}, default: {$ref: '#/components/responses/Error'}}}
  /approval-workflows/{id}/publish: {post: {operationId: publishApprovalWorkflow, tags: [Workflow], x-permission: workflow.manage, x-idempotency: required, x-phase: 6, x-supersedes: "/workflows/{id}/publish", description: "Published versions are immutable; running instances keep their version (doc 14).", parameters: [{$ref: '#/components/parameters/Id'}], responses: {'200': {$ref: '#/components/responses/Ok'}, default: {$ref: '#/components/responses/Error'}}}}
  /approval-tasks:                    {get:  {operationId: listApprovalTasks,   tags: [Workflow], x-permission: approval.view,     x-idempotency: n-a,      x-phase: 6, responses: {'200': {$ref: '#/components/responses/Ok'}, default: {$ref: '#/components/responses/Error'}}}}
  /approval-tasks/{id}/approve:       {post: {operationId: approveTask,         tags: [Workflow], x-permission: approval.approve,  x-idempotency: required, x-phase: 6, x-supersedes: "/approvals/{id}/approve", description: "SoD rules may forbid self-approval (doc 14, Gate F).", parameters: [{$ref: '#/components/parameters/Id'}], responses: {'200': {$ref: '#/components/responses/Ok'}, default: {$ref: '#/components/responses/Error'}}}}
  /approval-tasks/{id}/reject:        {post: {operationId: rejectTask,          tags: [Workflow], x-permission: approval.approve,  x-idempotency: required, x-phase: 6, x-supersedes: "/approvals/{id}/reject", parameters: [{$ref: '#/components/parameters/Id'}], responses: {'200': {$ref: '#/components/responses/Ok'}, default: {$ref: '#/components/responses/Error'}}}}
  /approval-tasks/{id}/request-changes: {post: {operationId: requestTaskChanges, tags: [Workflow], x-permission: approval.approve, x-idempotency: required, x-phase: 6, x-added-by: F-106, parameters: [{$ref: '#/components/parameters/Id'}], responses: {'200': {$ref: '#/components/responses/Ok'}, default: {$ref: '#/components/responses/Error'}}}}
  /approval-tasks/{id}/delegate:      {post: {operationId: delegateTask,        tags: [Workflow], x-permission: approval.delegate, x-idempotency: required, x-phase: 6, x-supersedes: "/approvals/{id}/delegate", description: "Delegation cannot violate SoD (doc 14).", parameters: [{$ref: '#/components/parameters/Id'}], responses: {'200': {$ref: '#/components/responses/Ok'}, default: {$ref: '#/components/responses/Error'}}}}
  /tasks:            {get:  {operationId: listTasks,    tags: [Workflow], x-permission: task.view,   x-idempotency: n-a,      x-phase: 6, x-added-by: F-106, description: "Unified inbox across 10 source types (doc 14).", responses: {'200': {$ref: '#/components/responses/Ok'}, default: {$ref: '#/components/responses/Error'}}}}
  /tasks/{id}/complete: {post: {operationId: completeTask, tags: [Workflow], x-permission: task.complete, x-idempotency: required, x-phase: 6, parameters: [{$ref: '#/components/parameters/Id'}], responses: {'200': {$ref: '#/components/responses/Ok'}, default: {$ref: '#/components/responses/Error'}}}}
...
components:
  securitySchemes:
    sessionBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        First-party session token. Carries a session ID checked against the `sessions` table on
        refresh, so revocation is immediate (ADR-0005). Tenant is bound to the session, not the
        credential.
    oauth2:
      type: oauth2
      description: Third-party applications (doc 15). Scopes map 1:1 to x-permission values.
      flows:
        authorizationCode:
          authorizationUrl: https://api.example.com/oauth/authorize
          tokenUrl: https://api.example.com/oauth/token
          scopes: {}
  parameters:
    Id:         {name: id,     in: path,  required: true,  schema: {type: string, format: uuid}}
...
  schemas:
...
    Error:
      type: object
      required: [code, message, correlation_id]
      properties:
        code:           {type: string, description: Stable machine-readable code. Never a stack trace or SQL (doc 15).}
        message:        {type: string}
        field_errors:   {type: array, items: {type: object, properties: {field: {type: string}, code: {type: string}, message: {type: string}}}}
        correlation_id: {type: string}
...
  responses:
    Ok:        {description: Success,  content: {application/json: {schema: {type: object}}}}
    Created:   {description: Created,  content: {application/json: {schema: {type: object}}}}
...
    Error:
      description: |
        Structured error with a stable code and correlation ID. Notable codes:
        DUPLICATE_SUSPECTED (409, vendor bill override), IN_PROGRESS (409, idempotency),
        PERIOD_CLOSED (422), APPROVAL_REQUIRED (403), UNBALANCED_JOURNAL (422),
        OVER_ALLOCATION (422), INSUFFICIENT_STOCK (422).
      content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}